fix: id de uint8_t a uint

+++ common/structures.h
@@ -32,7 +32,7 @@ enum EntityType {
 // Representación de una entidad en el mundo del juego
 struct Entity {
     EntityType type;
-    uint8_t id;
+    uint id;
     float x;
     float y;
 };
